fix nicktag issue

+++ WzComparerR2/CharaSimControl/ItemTooltipRender2.cs
@@ -1255,7 +1255,7 @@ namespace WzComparerR2.CharaSimControl
             }
             return checkString.Any(c => (c >= '\uAC00' && c <= '\uD7A3'));
         }
-        private bool TryGetNickResource(int nickTag, out Wz_Node resNode)
+        private bool TryGetNickResource(long nickTag, out Wz_Node resNode)
         {
             resNode = PluginBase.PluginManager.FindWz("UI/NameTag.img/nick/" + nickTag);
             return resNode != null;
